+++ JXCSystemServer/src/main/java/po/ItemPO.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class ItemPO implements Serializable {
 
 	public String item_name;
-	public int money;
+	public double money;
 	public String ps;
 	
 	
-	public ItemPO(String item_name,int money,String ps){
+	public ItemPO(String item_name,double money,String ps){
 		this.item_name=item_name;
 		this.money=money;
 		this.ps=ps;
@@ -18,7 +18,7 @@ public class ItemPO implements Serializable {
 	public String getItemName(){
 		return item_name;
 	}
-	public int getMoney(){
+	public double getMoney(){
 		return money;
 	}
 	public String getPs(){
